pyproject.toml
[tool.poetry]
name = "api-permissions"
version = "0.1.0"
description = ""
authors = ["joud-alnsour <alnsour.joud@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
install = "^1.3.5"
Markdown = "^3.3.7"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
